python/requirements.txt
nats-py==2.3.1
dataclass-wizard==0.22.2
prometheus-client==0.17.1
